+++ Assets/Prefabs/UI/SettingsMenu/Menus/EphysLinkMenu.prefab
@@ -1698,11 +1698,7 @@ MonoBehaviour:
   _connectionErrorText: {fileID: 642282321731354649}
   _manipulatorList: {fileID: 7994738088484906322}
   _manipulatorConnectionPanelPrefab: {fileID: 8830734852478451222, guid: 19384593a545a404bbe8bd298659df46, type: 3}
-  _automaticControlButton: {fileID: 1693939986521514926}
   _copilotButtonText: {fileID: 858957444264183286}
-  _destroyProbeEvent:
-    m_PersistentCalls:
-      m_Calls: []
 --- !u!114 &7858764217905313557
 MonoBehaviour:
   m_ObjectHideFlags: 0
