+++ Assets/Scripts/TrajectoryPlanner/UI/EphysCopilot/trajectoryplanner.ui.automaticmanipulaatorcontrol.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "trajectoryplanner.ui.automaticmanipulaatorcontrol",
+    "name": "trajectoryplanner.ui.ephyscopilot",
     "rootNamespace": "",
     "references": [
         "GUID:6055be8ebefd69e48b49212b09b47b2f",
